Condensed all score period modal function into score period modals file
