+++ src/pieces/discussion/menu/staff/manageScorePeriods/ScorePeriodModals.ts
@@ -1,8 +1,95 @@
-import { ButtonInteraction, Client, ModalSubmitInteraction } from "discord.js";
-import { getCourseByName, overwriteCourseDiscussionSpecs } from "../../../../../generalUtilities/CourseUtilities";
-import { CONFLICTING_DATES_MESSAGE, INVALID_INPUT_PREFIX, PERIOD_NUM_INPUT_ID, endDateActionRow, goalPointsActionRow, maxPointsActionRow, periodNumActionRow, startDateActionRow } from "./ModalComponents";
-import { NewPeriodData, checkAgainstCurrentPeriods, createHandlePeriodModal, handleIndexValidation, handlePeriodValidation, insertOnePeriod, overwritePeriods, sortPeriods, validatePeriodInput } from "./PeriodModalUtilities";
+import { ActionRowBuilder, ButtonInteraction, Client, ModalSubmitInteraction, TextInputBuilder, TextInputStyle } from "discord.js";
+import { DateTime } from "luxon";
+import { courseModel } from "../../../../../generalModels/Course";
+import { ScorePeriod } from "../../../../../generalModels/DiscussionScoring";
+import { DATABASE_ERROR_MESSAGE, getCourseByName, overwriteCourseDiscussionSpecs } from "../../../../../generalUtilities/CourseUtilities";
+import { sortPeriods } from "../../../../../generalUtilities/ScorePeriodUtilities";
+import { scoreAllThreads } from "../../../../../pieces/discussion/scoring/scoreFunctions";
+import { ModalInputHandler, createDiscussionModal } from "../../../../../pieces/menu/ModalUtilities";
+import { refreshManagePeriodsMenu, updateToManagePeriodsMenu } from "./ManageScorePeriodsMenu";
 
+// MODAL BEHAVIOR CONSTANTS
+export const DATE_STRING_FORMAT = "yyyy-MM-dd hh:mm:ss a";
+
+// MODAL NOTIFICATION CONSTANTS
+export const CONFLICTING_DATES_MESSAGE = "Score Period Has Overlap With Already Existing Score Period(s). Nothing Was Changed.";
+export const INVALID_INPUT_PREFIX = "Invalid Input Format. Nothing Was Changed\n**Reasons(s):**";
+export const INVALID_START_DATE_REASON = "\n- Invalid start date format. Input should be of the form: " + DATE_STRING_FORMAT.toUpperCase() + "M/PM and be before end date Ex: 1970-01-01 12:00:00 AM";
+export const INVALID_END_DATE_REASON = "\n- Invalid end date format. Input should be of the form: " + DATE_STRING_FORMAT.toUpperCase() + "M/PM and be after start date Ex: 2036-08-26 11:59:59 PM";
+export const INVALID_GOAL_POINTS_REASON = "\n- Invalid goal points. Input should be a non negative integer less than or equal to max points. Ex: 800";
+export const INVALID_MAX_POINTS_REASON = "\n- Invalid maximum points. Input should be a non negative integer greater than or equal to goal points. Ex: 1000";
+export const INVALID_INDEX_PERIOD_REASON = "\n- Invalid score period input. Please retry with a number in your menu."
+
+// INPUT FIELD CONSTANTS
+export const PERIOD_NUM_INPUT_ID = "discussion_score_period_input";
+const PERIOD_NUM_INPUT_LABEL = "score period #";
+const PERIOD_NUM_INPUT_PLACEHOLDER = "0";
+const PERIOD_NUM_INPUT_STYLE = TextInputStyle.Short;
+
+export const START_DATE_INPUT_ID = "discussion_score_period_start_input";
+const START_DATE_INPUT_LABEL = "start date/time: " + DATE_STRING_FORMAT.toUpperCase() + "M/PM";
+const START_DATE_INPUT_PLACEHOLDER = "1970-01-01 12:00:00 AM";
+const START_DATE_INPUT_STYLE = TextInputStyle.Short;
+
+export const END_DATE_INPUT_ID = "discussion_score_period_end_input";
+const END_DATE_INPUT_LABEL = "end date/time: " + DATE_STRING_FORMAT.toUpperCase() + "M/PM";
+const END_DATE_INPUT_PLACEHOLDER =  "2036-08-26 11:59:59 PM";
+const END_DATE_INPUT_STYLE = TextInputStyle.Short;
+
+export const GOAL_POINTS_INPUT_ID = "discussion_score_period_goal_input";
+const GOAL_POINTS_INPUT_LABEL = "goal points";
+const GOAL_POINTS_INPUT_PLACEHOLDER = "";
+const GOAL_POINTS_INPUT_STYLE = TextInputStyle.Short;
+
+const MAX_POINTS_INPUT_ID = "discussion_score_period_max_input";
+const MAX_POINTS_INPUT_LABEL = "max points";
+const MAX_POINTS_INPUT_PLACEHOLDER = "";
+const MAX_POINTS_INPUT_STYLE = TextInputStyle.Short;
+
+// INPUT FIELDS
+const periodNumInput = new TextInputBuilder({
+    customId: PERIOD_NUM_INPUT_ID,
+    label: PERIOD_NUM_INPUT_LABEL,
+    placeholder: PERIOD_NUM_INPUT_PLACEHOLDER,
+    style: PERIOD_NUM_INPUT_STYLE,
+})
+
+const startDateInput = new TextInputBuilder({
+    customId: START_DATE_INPUT_ID,
+    label: START_DATE_INPUT_LABEL,
+    placeholder: START_DATE_INPUT_PLACEHOLDER,
+    style: START_DATE_INPUT_STYLE,
+})
+
+const endDateInput = new TextInputBuilder({
+    customId: END_DATE_INPUT_ID,
+    label: END_DATE_INPUT_LABEL,
+    placeholder: END_DATE_INPUT_PLACEHOLDER,
+    style: END_DATE_INPUT_STYLE
+})
+
+const goalPointsInput = new TextInputBuilder({
+    customId: GOAL_POINTS_INPUT_ID,
+    label: GOAL_POINTS_INPUT_LABEL,
+    placeholder: GOAL_POINTS_INPUT_PLACEHOLDER,
+    style: GOAL_POINTS_INPUT_STYLE
+})
+
+const maxPointsInput = new TextInputBuilder({
+    customId: MAX_POINTS_INPUT_ID,
+    label: MAX_POINTS_INPUT_LABEL,
+    placeholder: MAX_POINTS_INPUT_PLACEHOLDER,
+    style: MAX_POINTS_INPUT_STYLE
+})
+
+// ACTION ROWS
+const periodNumActionRow = new ActionRowBuilder<TextInputBuilder>({components: [periodNumInput]});
+const startDateActionRow = new ActionRowBuilder<TextInputBuilder>({components: [startDateInput]});
+const endDateActionRow = new ActionRowBuilder<TextInputBuilder>({components: [endDateInput]});
+const goalPointsActionRow = new ActionRowBuilder<TextInputBuilder>({components: [goalPointsInput]});
+const maxPointsActionRow = new ActionRowBuilder<TextInputBuilder>({components: [maxPointsInput]});
+
+// ADD MODAL
 const ADD_MODAL_ID_PREFIX = "discussion_add_score_period_modal";
 const ADD_MODAL_TITLE_PREFIX = "Add Score Period To CISC ";
 const ADD_MODAL_SUCCESS_MESSAGE = "New Score Period Added!";
@@ -61,6 +148,7 @@ async function handleAddPeriodModal(client: Client, courseName: string, submitte
     return ADD_MODAL_SUCCESS_MESSAGE;
 }
 
+// EDIT MODAL
 const EDIT_MODAL_ID_PREFIX = "edit_score_period_modal";
 const EDIT_MODAL_TITLE_PREFIX = "Edit Score Period In CISC ";
 const EDIT_MODAL_SUCCESS_MESSAGE = "Score Period Successfully Edited";
@@ -123,6 +211,7 @@ async function handleEditPeriodModal(client: Client, courseName: string, submitt
     return EDIT_MODAL_SUCCESS_MESSAGE;
 }
 
+// DELETE MODAL
 const DELETE_MODAL_ID_PREFIX = "delete_score_period_modal"
 const DELETE_MODAL_TITLE_PREFIX = "Delete Score Period From CISC ";
 const DELETE_MODAL_SUCCESS_MESSAGE = "Score Period Was Successfully Removed!";
@@ -147,21 +236,152 @@ async function handleDeletePeriodModal(_client: Client, courseName: string, subm
 
     const fetchedCourse = await getCourseByName(courseName);
 
-    if(fetchedCourse && fetchedCourse.discussionSpecs !== null) {
+    if(!fetchedCourse || fetchedCourse.discussionSpecs === null)
+        return DATABASE_ERROR_MESSAGE
 
-        const reasonForFailure = handleIndexValidation(toDeleteIndex, fetchedCourse.discussionSpecs.scorePeriods.length)
+    const reasonForFailure = handleIndexValidation(toDeleteIndex, fetchedCourse.discussionSpecs.scorePeriods.length)
 
-        if(reasonForFailure !== "") {
-            return INVALID_INPUT_PREFIX + reasonForFailure;
-        }
-
-        fetchedCourse.discussionSpecs.scorePeriods.splice(toDeleteIndex - 1, 1);
-        
-        const deleteErrors = await overwriteCourseDiscussionSpecs(courseName, fetchedCourse.discussionSpecs);
-
-        if(deleteErrors !== "")
-            return deleteErrors
+    if(reasonForFailure !== "") {
+        return INVALID_INPUT_PREFIX + reasonForFailure;
     }
 
+    fetchedCourse.discussionSpecs.scorePeriods.splice(toDeleteIndex - 1, 1);
+        
+    const deleteErrors = await overwriteCourseDiscussionSpecs(courseName, fetchedCourse.discussionSpecs);
+
+    if(deleteErrors !== "")
+        return deleteErrors
+
     return DELETE_MODAL_SUCCESS_MESSAGE;
+}
+
+// UTILITY FUNCTIONS
+export async function createHandlePeriodModal(idPrefix: string, titlePrefix: string, courseName: string, triggerInteraction: ButtonInteraction, components: ActionRowBuilder<TextInputBuilder>[], modalInputHandler: ModalInputHandler) {
+    
+    updateToManagePeriodsMenu(courseName, triggerInteraction, false);
+
+    createDiscussionModal(idPrefix, titlePrefix, courseName, triggerInteraction, components, modalInputHandler, async () => {await refreshManagePeriodsMenu(courseName, triggerInteraction)})
+}
+
+interface PeriodValidationData {
+    startDate: Date | undefined, 
+    endDate: Date | undefined,  
+    goalPoints: number, 
+    maxPoints: number
+}
+
+function validatePeriodInput(submittedModal: ModalSubmitInteraction): PeriodValidationData {
+    
+    const startDateString = submittedModal.fields.getTextInputValue(START_DATE_INPUT_ID);
+    const startDateTime = DateTime.fromFormat(startDateString, DATE_STRING_FORMAT)
+    let startDate = startDateTime.toJSDate().getTime() ? startDateTime.toJSDate() : undefined;
+
+    const endDateString = submittedModal.fields.getTextInputValue(END_DATE_INPUT_ID);
+    const endDateTime = DateTime.fromFormat(endDateString, DATE_STRING_FORMAT)
+    let endDate = endDateTime.toJSDate().getTime() ? endDateTime.toJSDate() : undefined;
+
+    if(startDate && endDate && startDate.valueOf() >= endDate.valueOf()) {
+        startDate = undefined;
+        endDate = undefined;
+    }
+
+    let goalPoints = parseInt(submittedModal.fields.getTextInputValue(GOAL_POINTS_INPUT_ID));
+    let maxPoints = parseInt(submittedModal.fields.getTextInputValue(MAX_POINTS_INPUT_ID));
+    
+    if(goalPoints < 0)
+    goalPoints = NaN;
+
+    if(maxPoints < 0 )
+        maxPoints = NaN;
+    
+    if(goalPoints > maxPoints) {
+        goalPoints = NaN;
+        maxPoints = NaN;
+    }
+
+    return {startDate: startDate, endDate: endDate, goalPoints: goalPoints, maxPoints: maxPoints};
+}
+
+function handlePeriodValidation(modalData: PeriodValidationData): string { 
+    
+    let reasonsForFailure = "";
+    if(!modalData.startDate){
+        reasonsForFailure += INVALID_START_DATE_REASON;
+    }
+    if(!modalData.endDate){
+        reasonsForFailure += INVALID_END_DATE_REASON;
+    }
+    if(Number.isNaN(modalData.goalPoints)){
+        reasonsForFailure += INVALID_GOAL_POINTS_REASON;
+    }
+    if(Number.isNaN(modalData.maxPoints)){
+        reasonsForFailure += INVALID_MAX_POINTS_REASON;
+    }
+    return reasonsForFailure;
+}
+
+function handleIndexValidation(scorePeriodIndex: number, scorePeriodsLength: number): string {
+    
+    if( Number.isNaN(scorePeriodIndex) || scorePeriodIndex < 1 || (scorePeriodsLength && scorePeriodIndex > scorePeriodsLength) ) {
+        return INVALID_INDEX_PERIOD_REASON;
+    }
+    return "";
+}
+
+interface NewPeriodData {
+    start: Date, 
+    end: Date,  
+    goalPoints: number, 
+    maxPoints: number
+}
+
+async function checkAgainstCurrentPeriods(newScorePeriodData: NewPeriodData, currentScorePeriods: ScorePeriod[]): Promise<boolean> {
+
+    let hasConflict = false;
+    currentScorePeriods.forEach((scorePeriod) => {
+        if(scorePeriod.start.valueOf() <= newScorePeriodData.end.valueOf() && scorePeriod.end.valueOf() >= newScorePeriodData.start.valueOf()) {
+            hasConflict = true;
+        }
+    })
+    return hasConflict;
+}
+
+async function insertOnePeriod(client: Client, courseName: string, newScorePeriodData: NewPeriodData, scorePeriods: ScorePeriod[]): Promise<string> {
+
+    // TODO: After on the fly scoring is done, add listeners to undo and pause it while scoring is being done
+    scorePeriods = scorePeriods.sort((a, b) => { return a.start.valueOf() - b.start.valueOf() })
+
+    const course = await getCourseByName(courseName);
+
+    if(!course || !course.discussionSpecs || !course.channels.discussion)
+        return "database error"
+
+    course.discussionSpecs.scorePeriods.push({ ...newScorePeriodData, studentScores: new Map() });
+
+    // FIXME: This is currently broken because the rescored period isnt always the first/only one
+    // it was before the changes to rescoring done with post editing but now it isnt
+    // we likely just need to feed in discussion specs from the functions that call this instead of getting new ones from course since those dont have the removed/edited period
+    const newScorePeriods = await scoreAllThreads(client, course.channels.discussion, course.discussionSpecs, course.roles.staff)
+    
+    if(!newScorePeriods || newScorePeriods.length !== 1)
+        return "scoring error ocurred";
+
+    scorePeriods.push(newScorePeriods[0])
+    let insertErrors = await overwritePeriods(courseName, scorePeriods)
+    return insertErrors;
+}
+
+async function overwritePeriods(courseName: string, scorePeriods: ScorePeriod[]): Promise<string> {
+
+    try {
+        await courseModel.findOneAndUpdate( 
+            {name: courseName}, 
+            {"discussionSpecs.scorePeriods": scorePeriods}
+        )
+    }
+    catch(error: any) {
+        console.error(error);
+        return DATABASE_ERROR_MESSAGE;
+    }
+    return "";
 }
